# optimize cpu vec activations