… cmake/configure.cmake
… paddle/fluid/operators/math/CMakeLists.txt