sigmoid, tanh get about 100X boost, relu get about 1.28X

```
53: I0824 00:06:36.525028 32656 cpu_vec_test.cc:97] Vec size 1: refer takes: 0.313 us, tgt takes: 0.128
53: I0824 00:06:36.525446 32656 cpu_vec_test.cc:97] Vec size 1: refer takes: 0.303 us, tgt takes: 0.078
53: I0824 00:06:36.525846 32656 cpu_vec_test.cc:97] Vec size 1: refer takes: 0.308 us, tgt takes: 0.077
53: I0824 00:06:36.526240 32656 cpu_vec_test.cc:97] Vec size 1: refer takes: 0.302 us, tgt takes: 0.08
53: I0824 00:06:36.526939 32656 cpu_vec_test.cc:97] Vec size 2: refer takes: 0.604 us, tgt takes: 0.082
53: I0824 00:06:36.527643 32656 cpu_vec_test.cc:97] Vec size 2: refer takes: 0.608 us, tgt takes: 0.083
53: I0824 00:06:36.528340 32656 cpu_vec_test.cc:97] Vec size 2: refer takes: 0.6 us, tgt takes: 0.084
53: I0824 00:06:36.529042 32656 cpu_vec_test.cc:97] Vec size 2: refer takes: 0.606 us, tgt takes: 0.084
53: I0824 00:06:36.533676 32656 cpu_vec_test.cc:97] Vec size 15: refer takes: 4.484 us, tgt takes: 0.138
53: I0824 00:06:36.538293 32656 cpu_vec_test.cc:97] Vec size 15: refer takes: 4.478 us, tgt takes: 0.126
53: I0824 00:06:36.542914 32656 cpu_vec_test.cc:97] Vec size 15: refer takes: 4.482 us, tgt takes: 0.126
53: I0824 00:06:36.547530 32656 cpu_vec_test.cc:97] Vec size 15: refer takes: 4.478 us, tgt takes: 0.125
53: I0824 00:06:36.552438 32656 cpu_vec_test.cc:97] Vec size 16: refer takes: 4.776 us, tgt takes: 0.119
53: I0824 00:06:36.557361 32656 cpu_vec_test.cc:97] Vec size 16: refer takes: 4.78 us, tgt takes: 0.131
53: I0824 00:06:36.562271 32656 cpu_vec_test.cc:97] Vec size 16: refer takes: 4.778 us, tgt takes: 0.119
53: I0824 00:06:36.567190 32656 cpu_vec_test.cc:97] Vec size 16: refer takes: 4.778 us, tgt takes: 0.128
53: I0824 00:06:36.576900 32656 cpu_vec_test.cc:97] Vec size 32: refer takes: 9.548 us, tgt takes: 0.147
53: I0824 00:06:36.586609 32656 cpu_vec_test.cc:97] Vec size 32: refer takes: 9.551 us, tgt takes: 0.145
53: I0824 00:06:36.596328 32656 cpu_vec_test.cc:97] Vec size 32: refer takes: 9.558 us, tgt takes: 0.148
53: I0824 00:06:36.606048 32656 cpu_vec_test.cc:97] Vec size 32: refer takes: 9.555 us, tgt takes: 0.151
53: I0824 00:06:36.645669 32656 cpu_vec_test.cc:97] Vec size 128: refer takes: 38.216 us, tgt takes: 1.38
53: I0824 00:06:36.684492 32656 cpu_vec_test.cc:97] Vec size 128: refer takes: 38.193 us, tgt takes: 0.585
53: I0824 00:06:36.723379 32656 cpu_vec_test.cc:97] Vec size 128: refer takes: 38.275 us, tgt takes: 0.589
53: I0824 00:06:36.762274 32656 cpu_vec_test.cc:97] Vec size 128: refer takes: 38.282 us, tgt takes: 0.587
53: I0824 00:06:36.822919 32656 cpu_vec_test.cc:97] Vec size 200: refer takes: 59.797 us, tgt takes: 0.817
53: I0824 00:06:36.883411 32656 cpu_vec_test.cc:97] Vec size 200: refer takes: 59.658 us, tgt takes: 0.801
53: I0824 00:06:36.944057 32656 cpu_vec_test.cc:97] Vec size 200: refer takes: 59.818 us, tgt takes: 0.805
53: I0824 00:06:37.004541 32656 cpu_vec_test.cc:97] Vec size 200: refer takes: 59.658 us, tgt takes: 0.801
53: I0824 00:06:37.159176 32656 cpu_vec_test.cc:97] Vec size 512: refer takes: 152.757 us, tgt takes: 1.83
53: I0824 00:06:37.313973 32656 cpu_vec_test.cc:97] Vec size 512: refer takes: 152.935 us, tgt takes: 1.806
53: I0824 00:06:37.468631 32656 cpu_vec_test.cc:97] Vec size 512: refer takes: 152.819 us, tgt takes: 1.801
53: I0824 00:06:37.623292 32656 cpu_vec_test.cc:97] Vec size 512: refer takes: 152.82 us, tgt takes: 1.808
53: I0824 00:06:37.624549 32656 cpu_vec_test.cc:97] Vec size 30: refer takes: 0.877 us, tgt takes: 0.355
53: [       OK ] CpuVecTest.sigmoid (1100 ms)
53: [ RUN      ] CpuVecTest.tanh
53: I0824 10:03:35.012017 21516 cpu_vec_test.cc:97] Vec size 1: refer takes: 0.305 us, tgt takes: 0.104
53: I0824 10:03:35.012425 21516 cpu_vec_test.cc:97] Vec size 1: refer takes: 0.306 us, tgt takes: 0.087
53: I0824 10:03:35.012832 21516 cpu_vec_test.cc:97] Vec size 1: refer takes: 0.31 us, tgt takes: 0.086
53: I0824 10:03:35.013236 21516 cpu_vec_test.cc:97] Vec size 1: refer takes: 0.305 us, tgt takes: 0.087
53: I0824 10:03:35.013949 21516 cpu_vec_test.cc:97] Vec size 2: refer takes: 0.61 us, tgt takes: 0.091
53: I0824 10:03:35.014663 21516 cpu_vec_test.cc:97] Vec size 2: refer takes: 0.607 us, tgt takes: 0.095
53: I0824 10:03:35.015391 21516 cpu_vec_test.cc:97] Vec size 2: refer takes: 0.61 us, tgt takes: 0.106
53: I0824 10:03:35.016101 21516 cpu_vec_test.cc:97] Vec size 2: refer takes: 0.606 us, tgt takes: 0.093
53: I0824 10:03:35.020844 21516 cpu_vec_test.cc:97] Vec size 15: refer takes: 4.538 us, tgt takes: 0.192
53: I0824 10:03:35.025533 21516 cpu_vec_test.cc:97] Vec size 15: refer takes: 4.523 us, tgt takes: 0.153
53: I0824 10:03:35.030246 21516 cpu_vec_test.cc:97] Vec size 15: refer takes: 4.543 us, tgt takes: 0.156
53: I0824 10:03:35.034942 21516 cpu_vec_test.cc:97] Vec size 15: refer takes: 4.536 us, tgt takes: 0.148
53: I0824 10:03:35.039986 21516 cpu_vec_test.cc:97] Vec size 16: refer takes: 4.885 us, tgt takes: 0.133
53: I0824 10:03:35.045017 21516 cpu_vec_test.cc:97] Vec size 16: refer takes: 4.85 us, tgt takes: 0.13
53: I0824 10:03:35.049998 21516 cpu_vec_test.cc:97] Vec size 16: refer takes: 4.838 us, tgt takes: 0.127
53: I0824 10:03:35.054986 21516 cpu_vec_test.cc:97] Vec size 16: refer takes: 4.849 us, tgt takes: 0.125
53: I0824 10:03:35.064895 21516 cpu_vec_test.cc:97] Vec size 32: refer takes: 9.72 us, tgt takes: 0.174
53: I0824 10:03:35.074784 21516 cpu_vec_test.cc:97] Vec size 32: refer takes: 9.705 us, tgt takes: 0.166
53: I0824 10:03:35.084617 21516 cpu_vec_test.cc:97] Vec size 32: refer takes: 9.654 us, tgt takes: 0.166
53: I0824 10:03:35.094493 21516 cpu_vec_test.cc:97] Vec size 32: refer takes: 9.696 us, tgt takes: 0.164
53: I0824 10:03:35.133867 21516 cpu_vec_test.cc:97] Vec size 128: refer takes: 38.732 us, tgt takes: 0.623
53: I0824 10:03:35.173352 21516 cpu_vec_test.cc:97] Vec size 128: refer takes: 38.809 us, tgt takes: 0.654
53: I0824 10:03:35.212726 21516 cpu_vec_test.cc:97] Vec size 128: refer takes: 38.694 us, tgt takes: 0.653
53: I0824 10:03:35.252128 21516 cpu_vec_test.cc:97] Vec size 128: refer takes: 38.751 us, tgt takes: 0.626
53: I0824 10:03:35.313661 21516 cpu_vec_test.cc:97] Vec size 200: refer takes: 60.628 us, tgt takes: 0.876
53: I0824 10:03:35.375099 21516 cpu_vec_test.cc:97] Vec size 200: refer takes: 60.54 us, tgt takes: 0.87
53: I0824 10:03:35.436566 21516 cpu_vec_test.cc:97] Vec size 200: refer takes: 60.539 us, tgt takes: 0.904
53: I0824 10:03:35.497973 21516 cpu_vec_test.cc:97] Vec size 200: refer takes: 60.511 us, tgt takes: 0.872
53: I0824 10:03:35.655630 21516 cpu_vec_test.cc:97] Vec size 512: refer takes: 155.586 us, tgt takes: 2.032
53: I0824 10:03:35.812531 21516 cpu_vec_test.cc:97] Vec size 512: refer takes: 154.841 us, tgt takes: 2.016
53: I0824 10:03:35.969157 21516 cpu_vec_test.cc:97] Vec size 512: refer takes: 154.664 us, tgt takes: 1.92
53: I0824 10:03:36.126001 21516 cpu_vec_test.cc:97] Vec size 512: refer takes: 154.816 us, tgt takes: 1.994
53: I0824 10:03:36.127301 21516 cpu_vec_test.cc:97] Vec size 30: refer takes: 0.911 us, tgt takes: 0.364
53: [       OK ] CpuVecTest.tanh (1101 ms)
53: [ RUN      ] CpuVecTest.relu
53: I0824 00:06:38.725839 32656 cpu_vec_test.cc:97] Vec size 1: refer takes: 0.006 us, tgt takes: 0.007
53: I0824 00:06:38.725864 32656 cpu_vec_test.cc:97] Vec size 1: refer takes: 0.006 us, tgt takes: 0.007
53: I0824 00:06:38.725888 32656 cpu_vec_test.cc:97] Vec size 1: refer takes: 0.006 us, tgt takes: 0.007
53: I0824 00:06:38.725913 32656 cpu_vec_test.cc:97] Vec size 1: refer takes: 0.006 us, tgt takes: 0.006
53: I0824 00:06:38.725936 32656 cpu_vec_test.cc:97] Vec size 2: refer takes: 0.008 us, tgt takes: 0.007
53: I0824 00:06:38.725961 32656 cpu_vec_test.cc:97] Vec size 2: refer takes: 0.007 us, tgt takes: 0.008
53: I0824 00:06:38.725987 32656 cpu_vec_test.cc:97] Vec size 2: refer takes: 0.007 us, tgt takes: 0.007
53: I0824 00:06:38.726012 32656 cpu_vec_test.cc:97] Vec size 2: refer takes: 0.007 us, tgt takes: 0.007
53: I0824 00:06:38.726049 32656 cpu_vec_test.cc:97] Vec size 15: refer takes: 0.013 us, tgt takes: 0.012
53: I0824 00:06:38.726083 32656 cpu_vec_test.cc:97] Vec size 15: refer takes: 0.013 us, tgt takes: 0.009
53: I0824 00:06:38.726114 32656 cpu_vec_test.cc:97] Vec size 15: refer takes: 0.013 us, tgt takes: 0.007
53: I0824 00:06:38.726145 32656 cpu_vec_test.cc:97] Vec size 15: refer takes: 0.013 us, tgt takes: 0.007
53: I0824 00:06:38.726174 32656 cpu_vec_test.cc:97] Vec size 16: refer takes: 0.009 us, tgt takes: 0.009
53: I0824 00:06:38.726202 32656 cpu_vec_test.cc:97] Vec size 16: refer takes: 0.009 us, tgt takes: 0.007
53: I0824 00:06:38.726230 32656 cpu_vec_test.cc:97] Vec size 16: refer takes: 0.009 us, tgt takes: 0.008
53: I0824 00:06:38.726258 32656 cpu_vec_test.cc:97] Vec size 16: refer takes: 0.01 us, tgt takes: 0.007
53: I0824 00:06:38.726294 32656 cpu_vec_test.cc:97] Vec size 32: refer takes: 0.012 us, tgt takes: 0.012
53: I0824 00:06:38.726328 32656 cpu_vec_test.cc:97] Vec size 32: refer takes: 0.013 us, tgt takes: 0.009
53: I0824 00:06:38.726362 32656 cpu_vec_test.cc:97] Vec size 32: refer takes: 0.012 us, tgt takes: 0.01
53: I0824 00:06:38.726397 32656 cpu_vec_test.cc:97] Vec size 32: refer takes: 0.012 us, tgt takes: 0.01
53: I0824 00:06:38.726469 32656 cpu_vec_test.cc:97] Vec size 128: refer takes: 0.032 us, tgt takes: 0.027
53: I0824 00:06:38.726541 32656 cpu_vec_test.cc:97] Vec size 128: refer takes: 0.032 us, tgt takes: 0.025
53: I0824 00:06:38.726615 32656 cpu_vec_test.cc:97] Vec size 128: refer takes: 0.033 us, tgt takes: 0.026
53: I0824 00:06:38.726691 32656 cpu_vec_test.cc:97] Vec size 128: refer takes: 0.032 us, tgt takes: 0.025
53: I0824 00:06:38.726794 32656 cpu_vec_test.cc:97] Vec size 200: refer takes: 0.048 us, tgt takes: 0.037
53: I0824 00:06:38.726894 32656 cpu_vec_test.cc:97] Vec size 200: refer takes: 0.047 us, tgt takes: 0.036
53: I0824 00:06:38.726996 32656 cpu_vec_test.cc:97] Vec size 200: refer takes: 0.047 us, tgt takes: 0.037
53: I0824 00:06:38.727097 32656 cpu_vec_test.cc:97] Vec size 200: refer takes: 0.047 us, tgt takes: 0.036
53: I0824 00:06:38.727339 32656 cpu_vec_test.cc:97] Vec size 512: refer takes: 0.121 us, tgt takes: 0.095
53: I0824 00:06:38.727582 32656 cpu_vec_test.cc:97] Vec size 512: refer takes: 0.122 us, tgt takes: 0.094
53: I0824 00:06:38.727830 32656 cpu_vec_test.cc:97] Vec size 512: refer takes: 0.121 us, tgt takes: 0.099
53: I0824 00:06:38.728073 32656 cpu_vec_test.cc:97] Vec size 512: refer takes: 0.121 us, tgt takes: 0.094
53: I0824 00:06:38.728126 32656 cpu_vec_test.cc:97] Vec size 30: refer takes: 0.017 us, tgt takes: 0.017
```