@tensor-tang 2018-08-24
```
	[ 57%] Linking CXX shared library /paddle/build/third_party/anakin/src/extern_anakin/output/libanakin.so[ 57%] Lin
[02:52:09]	nvcc fatal   : redefinition of argument 'machine'
[02:52:09]	CMake Error at concat_generated_concat.cu.o.cmake:203 (message):
[02:52:09]	  Error generating
[02:52:09]	  /paddle/build/paddle/fluid/operators/math/CMakeFiles/concat.dir//./concat_generated_concat.cu.o
[02:52:09]	
[02:52:09]	
[02:52:09]	nvcc fatal   : redefinition of argument 'machine'
[02:52:09]	CMake Error at lstm_compute_generated_lstm_compute.cu.o.cmake:203 (message):
[02:52:09]	  Error generating
[02:52:09]	  /paddle/build/paddle/fluid/operators/math/CMakeFiles/lstm_compute.dir//./lstm_compute_generated_lstm_compute.cu.o
[02:52:09]	
[02:52:09]	
[02:52:09]	make[2]: *** [paddle/fluid/operators/math/CMakeFiles/concat.dir/concat_generated_concat.cu.o] Error 1
[02:52:09]	nvcc fatal   : redefinition of argument 'machine'
```