@tensor-tang 2018-08-24
the data [here](https://github.com/PaddlePaddle/Paddle/pull/12878#issuecomment-415049218) speedup from 61.9us to ~~22.758~~ 20.723 us